requirements.txt
#requirements.txt - auto-generated
tensorflow>=2.11.0
transformers>=4.36.0
datasets>=2.16.0
huggingface_hub>=0.20.0
tensorflow_hub>=0.13.0
numpy>=1.22.0
pillow>=9.3.0
scikit-learn>=1.1.3
matplotlib>=3.6.2
jupyterlab>=3.5.0
streamlit